renaming the InequalityPreviews.kt file to InequalityLightPreviews.kt and its refactoring
